# Optimize `compute_by()` for the "no groups" case