~25% less overhead in `mutate()` now

``` r
library(dplyr)

df <- tibble(x = c(1, 2, 3, 4))

bench::mark(mutate(df, y = x + 1), iterations = 10000)

# Main
#> # A tibble: 1 × 6
#>   expression                 min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>            <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 mutate(df, y = x + 1)   1.72ms   1.96ms      483.    2.76MB     11.1

# This PR
#> # A tibble: 1 × 6
#>   expression                 min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>            <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 mutate(df, y = x + 1)    986µs   1.42ms      678.    2.31MB     10.1
```